✏️ [Rename] 설정, 프로필 폴더 분리
